moved component
Added weaponStats and moved component to Tooltip to keep CharacterItemGrid less cluttered
