+++ src/components/Tooltip.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { createPortal } from "react-dom";
+import { WowItem } from "../pages/View";
 
 interface TooltipProps {
   children: React.ReactElement;
@@ -77,4 +78,128 @@ const Tooltip: React.FC<TooltipProps> = ({
   );
 };
 
-export default Tooltip;
+interface TooltipWowItemProps {
+  item: WowItem;
+  fallback: string;
+}
+
+const TooltipWowItem: React.FC<TooltipWowItemProps> = ({ item, fallback }) => {
+  const qualityBorderColor: { [key: string]: string } = {
+    Poor: "border-gray-500",
+    Common: "border-white",
+    Uncommon: "border-green-500",
+    Rare: "border-blue-500",
+    Epic: "border-purple-500",
+    Legendary: "border-orange-500",
+  };
+
+  const qualityNameColor: { [key: string]: string } = {
+    Poor: "text-gray-500",
+    Common: "border-white",
+    Uncommon: "text-green-500",
+    Rare: "text-blue-500",
+    Epic: "text-purple-500",
+    Legendary: "text-orange-500",
+  };
+
+  const borderColor = item
+    ? qualityBorderColor[item.quality] || "border-transparent"
+    : "border-transparent";
+  const nameColor = item
+    ? qualityNameColor[item.quality] || "text-white"
+    : "text-white";
+
+  interface PriceProps {
+    header: string;
+    gold?: string;
+    silver?: string;
+    copper?: string;
+  }
+
+  function formatSellPrice(price: PriceProps) {
+    return (
+      <div className="flex gap-x-1.5">
+        {price.header}
+        {price.gold && price.gold !== "0" && price.gold !== "" && (
+          <span className="flex items-center">
+            {price.gold}
+            <img
+              src="/icons/currency/gold.webp"
+              alt="gold"
+              className="w-4 h-4"
+            />
+          </span>
+        )}
+        {price.silver && price.silver !== "0" && price.silver !== "" && (
+          <div className="flex items-center">
+            {price.silver}
+            <img
+              src="/icons/currency/silver.webp"
+              alt="silver"
+              className="w-4 h-4"
+            />
+          </div>
+        )}
+        {price.copper && price.copper !== "0" && price.copper !== "" && (
+          <div className="flex items-center">
+            {price.copper}
+            <img
+              src="/icons/currency/copper.webp"
+              alt="copper"
+              className="w-4 h-4"
+            />
+          </div>
+        )}
+      </div>
+    );
+  }
+
+  return (
+    <Tooltip
+      offset={{ x: 10, y: -40 }}
+      content={
+        <div className="flex flex-col bg-black text-white py-1 px-2 rounded-lg max-w-sm">
+          <span className={`block ${nameColor}`}>{item.name}</span>
+          <span className="text-yellow-400">Item level {item.level}</span>
+          <div className="flex justify-between">
+            <span>{item.slot}</span>
+            {item.itemSubclass !== "Miscellaneous" && (
+              <span>{item.itemSubclass}</span>
+            )}
+          </div>
+          {item.weaponStats && (
+            <div className="flex flex-col">
+              <div className="flex justify-between space-x-6">
+                <span>{item.weaponStats.damage}</span>
+                <span>{item.weaponStats.attackSpeed}</span>
+              </div>
+              <span>{item.weaponStats.dps}</span>
+            </div>
+          )}
+          <span>{item.armor}</span>
+          <div className="flex flex-col">
+            {item.stats.map((stat, index) => (
+              <div key={index}>{stat}</div>
+            ))}
+          </div>
+          <span>{item.durability}</span>
+          <span>Requires Level {item.requiredLevel}</span>
+          <div className="flex flex-col text-green-500">
+            {item.spells.map((spell, index) => (
+              <div key={index}>{spell}</div>
+            ))}
+          </div>
+          <span>{formatSellPrice(item.sellPrice)}</span>
+        </div>
+      }
+    >
+      <img
+        src={item ? item.icon : `/icons/gear/${fallback}.webp`}
+        alt={item ? item.name : fallback}
+        className={`w-[56px] h-[56px] object-contain rounded-lg shadow-sm border-4 ${borderColor}`}
+      />
+    </Tooltip>
+  );
+};
+
+export default TooltipWowItem;
